Change: IsValid replaces null checks on pinned asset
Elements in pinned lists should NOT be null, not only due to serialization, but also because list elements should never be set to null. Elements are either added or removed. Because of this, it is safe to use `IsValid` to check data rather than checking if the pinned asset instance is null.